mv meta exports to subdirs
