+++ tools/meta/src/getAllPackages/index.ts
@@ -1,0 +1,1 @@
+export { getAllPackageNames, getAllPackages } from './getAllPackages';
