+++ tools/meta/src/getPackageManager/getPackageManager.spec.ts
