+++ tools/meta/src/getPackageManager/getPackageManager.ts
