+++ tools/meta/src/getPackageManager/index.ts
@@ -1,0 +1,4 @@
+export {
+  getPackageManager,
+  type SupportedPackageManager,
+} from './getPackageManager';
