+++ tools/meta/src/getPackageName/index.ts
@@ -1,0 +1,1 @@
+export { getPackageName } from './getPackageName';
